pyproject.toml
[tool.poetry]
name = "gnosis-safe-slackbot"
version = "0.1.0"
description = "Slack Bot that will periodically fetch queued transactions in a Gnosis Safe and post them in slack"
authors = ["verumlotus"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
